requirements/runtime.txt
pydantic
PyYAML
python-dotenv
jinja2
click
prompt_toolkit
croniter
watchdog
fastjsonschema
cachetools
psutil
screeninfo
pywin32
dxcam
windows-capture
av==17.0.0
opencv-python
numpy
pillow
